Fix some documentation
