+++ pymeasure/adapters/fake_scpi.py
@@ -52,29 +52,15 @@ class FakeScpiAdapter(Adapter):
       numerical, but only as strings. (TODO: Implement other IEEE 488.2 integer
       formats).
 
-    .. code-block:: python
+    Example usage:
 
-        a = FakeScpiAdapter(default_value=111, MEAS=33)
-        assert a.read() == ""
-        a.write("ABCD?")
-        assert a.read() == "111\n"  # default value
-        a.write("EFGH 4")
-        assert a.ask("EFGH?") == "4\n"
-        assert a.ask("MEAS?") == "33\n"
-
-        def handler(self, args=tuple(), query=False):
-            if query:
-                if args:
-                    return str(sum(args))
-                else:
-                    return '0'
-
-        a.set_handler("SUM", handler)
-        assert a.ask("SUM? 1,2,3,4") == "10"
+    .. literalinclude:: /api/example-FakeScpiAdapter.py
+        :language: python
 
     :param default_value: Default value to return for queries to keys that have
         not been set. If None, then queries for unset keys will throw a
         KeyError.
+
     :param kwargs: Keyword arguments of the form ``CMDD=value``, where CMDD is
         the SCPI command mnemonic (key) and ``value`` is the initial value for
         queries to that key (i.e. ``"CMDD?"``) or, if a callable, the handler
